@@ -1,9 +1,6 @@
 ﻿using RevXApi.Library.Models;
-using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace RevXApi.Library.DataAccess
 {
@@ -21,6 +18,11 @@ namespace RevXApi.Library.DataAccess
 			return _sql.LoadData<BillingStatusModel>("dbo.spBillingStatus_GetAll", "RevXData");
 		}
 
+		public List<BillingStatusModel> GetEnabled()
+		{
+			return _sql.LoadData<BillingStatusModel>("dbo.spBillingStatus_GetEnabled", "RevXData");
+		}
+
 		public BillingStatusModel GetById(int id)
 		{
 			return _sql.LoadData<BillingStatusModel, dynamic>("dbo.spBillingStatus_GetById", new { Id = id }, "RevXData").FirstOrDefault();
@@ -29,6 +31,11 @@ namespace RevXApi.Library.DataAccess
 		public void AddBillingStatus(BillingStatusModel model)
 		{
 			_sql.SaveData("dbo.spBillingStatus_Insert", model, "RevXData");
+		}
+
+		public void EditBillingStatus(BillingStatusModel model)
+		{
+			_sql.SaveData("dbo.spBillingStatus_Update", model, "RevXData");
 		}
 	}
 }
