@@ -5,11 +5,11 @@ namespace RevXApi.Library.DataAccess
 {
 	public interface ISessionData
 	{
-		List<SessionModel> GetAllSessions();
-		SessionModel GetById(int id);
+		List<SessionModel> GetAllSessions(string userId);
+		SessionModel GetById(int id, string userId);
 		List<SessionModel> GetByBillingStatus(BillingStatusModel billingStatus);
-		void SaveSession(SessionModel model);
+		int SaveSession(SessionModel model);
 		void EditSession(SessionModel model);
-		void DeleteSession(int id);
+		void DeleteSession(int id, string userId);
 	}
 }
