@@ -1,8 +1,13 @@
-﻿using RevXApi.Library.Models;
+﻿using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Logging;
+using RazorEngine;
+using RazorEngine.Templating;
+using RevXApi.Library.Models;
+using SelectPdf;
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace RevXApi.Library.DataAccess
@@ -11,42 +16,51 @@ namespace RevXApi.Library.DataAccess
 	{
 		private readonly ISqlDataAccess _sql;
 		private readonly ISessionData _sessionData;
+		private readonly IUserData _userData;
+		private readonly IHourlyRateData _rateData;
+		private readonly ILogger<InvoiceData> _logger;
+		private readonly IConfiguration _config;
+		private readonly string _documentLocationRoot;
 
-		public InvoiceData(ISqlDataAccess sql, ISessionData sessionData)
+		public InvoiceData(ISqlDataAccess sql, ISessionData sessionData, IUserData userData, IHourlyRateData rateData, ILogger<InvoiceData> logger, IConfiguration config)
 		{
 			_sql = sql;
 			_sessionData = sessionData;
+			_userData = userData;
+			_rateData = rateData;
+			_logger = logger;
+			_config = config;
+			_documentLocationRoot = _config[ "EmailConfig:DocumentTemplateLocationRoot" ];
 		}
 
-		public void SaveInvoice(InvoiceModel invoice)
+
+		public List<InvoiceModel> GetAll(string userId)
+		{
+			return _sql.LoadData<InvoiceModel, dynamic>("dbo.spInvoice_GetAll", new { userId }, "RevXData");
+		}
+
+		public int SaveInvoice(InvoiceModel invoice)
 		{
 			try
 			{
 				_sql.StartTransaction("RevXData");
 
 				// Save the Invoice
-				_sql.SaveDataInTransaction("dbo.spInvoice_Insert", new { invoice.Id, invoice.InvoiceDate, invoice.TotalHours });
+				invoice.Id = _sql.SaveDataInTransactionWithResult<int, InvoiceDBModel>("dbo.spInvoice_Insert", new() { InvoiceDate = invoice.InvoiceDate, StartDate = invoice.StartDate, EndDate = invoice.EndDate, TotalHours = invoice.TotalHours, Rate = invoice.Rate, Total = invoice.Total, UserId = invoice.UserId, ProviderId = invoice.ProviderId }).FirstOrDefault();
 
-				// Get back the Id of this invoice
-				invoice.Id = _sql.LoadDataInTransaction<int, dynamic>("dbo.spInvoice_Lookup", new { invoice.InvoiceDate, invoice.TotalHours }).FirstOrDefault();
-				
+				// 0 is Default for the int type
 				if (invoice.Id > 0)
 				{
 					foreach (var id in invoice.SessionIds)
 					{
-						var detail = new InvoiceDetailModel() { InvoiceId = invoice.Id, SessionId = id };
-
-						// Save the Detail to database
-						_sql.SaveDataInTransaction("dbo.spInvoiceDetail_Insert", detail);
-
-						// Update the sessions billing status to invoiced
-						// TODO - the status id should be looked up, not manually typed!!!!
-						_sql.SaveDataInTransaction("dbo.spSession_EditBillingStatus", new { Id = id, StatusId = 2 });
+						int affected = _sql.ExecuteCommandInTransaction<int>($"UPDATE Session SET BillingStatusId = {2}, InvoiceId = {invoice.Id} WHERE Id = {id} AND UserId = '{invoice.UserId}'");
 					}
 
 					_sql.CommitTransaction();
+					return invoice.Id;
 				}
-				else throw new Exception();
+				else
+					throw new Exception("New invoice id was 0.");
 			}
 			catch
 			{
@@ -58,28 +72,132 @@ namespace RevXApi.Library.DataAccess
 		public InvoiceEmailModel PrepareEmailModel(InvoiceModel invoice)
 		{
 			// Fill in the basics
-			InvoiceEmailModel output = new() { InvoiceDate = invoice.InvoiceDate, TotalHours = invoice.TotalHours };
+			InvoiceEmailModel output = new() { InvoiceDate = invoice.InvoiceDate, TotalHours = invoice.TotalHours, Rate = invoice.Rate, Signature = invoice.Signature, UserId = invoice.UserId };
+
+			// Get the Name of employee
+			UserModel employee = _userData.GetUserById(invoice.UserId);
+			string employeeName = $"{employee.FirstName} {employee.LastName}";
+			output.FullName = employeeName;
 
 			// Get the full session models, and change to email model
 			output.InvoiceSessions = new();
 			foreach (var item in invoice.SessionIds)
 			{
-				var temp = _sessionData.GetById(item);
+				var temp = _sessionData.GetById(item, invoice.UserId);
 				SessionEmailModel model = new()
 				{
 					Student = $"{temp.Student.FirstName} {temp.Student.LastName}",
 					Date = temp.Date,
-					StartTime = Convert24HourTo12Hour(temp.StartTime),
-					EndTime = Convert24HourTo12Hour(temp.EndTime),
+					StartTime = TimeSpan.Parse(temp.StartTime),
+					EndTime = TimeSpan.Parse(temp.EndTime),
 					Notes = temp.Notes
 				};
 				output.InvoiceSessions.Add(model);
 			}
 
 			// Calculate the invoicing period
+			output.InvoicePeriod = CalculateInvoicePeriod(output);
+
+			return output;
+
+		}
+
+		public List<InvoiceModel> GenerateInvoicesFromSessions(List<SessionDbModel> sessions, string userId)
+		{
+			_logger.LogInformation(message: $"I got some sessions for {userId} to generate invoices", userId);
+			Dictionary<string, List<SessionDbModel>> keyValuePairs = new();
+			foreach (var session in sessions)
+			{
+				int month = session.Date.Month;
+				int year = session.Date.Year;
+				if (keyValuePairs.ContainsKey($"{month}-{year}-{session.ProviderId}"))
+				{
+					keyValuePairs.GetValueOrDefault($"{month}-{year}-{session.ProviderId}").Add(session);
+				}
+				else
+				{
+					keyValuePairs.Add($"{month}-{year}-{session.ProviderId}", new List<SessionDbModel> {session});
+				}
+			}
+			List<InvoiceModel> output = new();
+			foreach (List<SessionDbModel> month in keyValuePairs.Values)
+			{
+				month.Sort((s, x) => s.Date < x.Date ? -1 : 1);
+				var invoice = new InvoiceModel() { UserId = userId };
+				invoice.SessionIds = month.Select(x => x.Id).ToList();
+				invoice.InvoiceDate = DateTime.Now;
+				invoice.StartDate = month.First().Date;
+				invoice.EndDate = month.Last().Date;
+				invoice.Rate = _rateData.GetByDate(invoice.StartDate, invoice.UserId, month.First().ProviderId).Rate;
+				invoice.ProviderId = month.First().ProviderId;
+				invoice.TotalHours = month.Select(s => (s.EndTime - s.StartTime).TotalHours).Sum();
+				invoice.Total = invoice.Rate * invoice.TotalHours;
+				output.Add(invoice);
+			}
+			return output;
+		}
+
+		public byte[] GetDocument(int id, string userId)
+		{
+			InvoiceEmailModel model = _sql.LoadData<InvoiceEmailModel, dynamic>("spInvoice_Lookup", new { InvoiceId = id, UserId = userId }, "RevXData").FirstOrDefault();
+			if (model == null)
+			{
+				return new byte[0];
+			}
+			model.InvoiceSessions = _sql.Query<SessionEmailModel>($"SELECT se.Id, CONCAT(st.FirstName, ' ', st.LastName) as Student, se.[Date], se.StartTime, se.EndTime, se.Notes FROM Session se JOIN Student st ON se.StudentId = st.Id WHERE se.InvoiceId = {id} AND se.UserId = '{userId}'", "RevXData");
+			model.InvoicePeriod = CalculateInvoicePeriod(model);
+			UserModel user = _userData.GetUserById(userId);
+			model.FullName = user.FirstName + " " + user.LastName;
+			string template = File.ReadAllText(_documentLocationRoot + "/InvoiceDocument.cshtml");
+			string result = Engine.Razor.RunCompile(template, DateTime.Now.ToString(), null, model);
+			// instantiate the html to pdf converter
+			HtmlToPdf converter = new();
+			converter.Options.MarginBottom = 30;
+			converter.Options.MarginTop = 30;
+			// convert the url to pdf
+			PdfDocument doc = converter.ConvertHtmlString(result);
+
+			// save pdf document
+			var res = doc.Save();
+
+			// close pdf document
+			return res;
+		}
+
+
+		public byte[] GetDocument(int id, string userId, string signature)
+		{
+			InvoiceEmailModel model = _sql.LoadData<InvoiceEmailModel, dynamic>("spInvoice_Lookup", new { InvoiceId = id, UserId = userId }, "RevXData").FirstOrDefault();
+			_logger.LogInformation("model was set to {0}", model.Id);
+			model.InvoiceSessions = _sql.Query<SessionEmailModel>($"SELECT se.Id, CONCAT(st.FirstName, ' ', st.LastName) as Student, se.[Date], se.StartTime, se.EndTime, se.Notes FROM Session se JOIN Student st ON se.StudentId = st.Id WHERE se.InvoiceId = {id} AND se.UserId = '{userId}'", "RevXData");
+			model.InvoicePeriod = CalculateInvoicePeriod(model);
+			model.Signature = signature;
+			UserModel user = _userData.GetUserById(userId);
+			model.FullName = user.FirstName + " " + user.LastName;
+			_logger.LogInformation("Trying to get the template from {0}", _documentLocationRoot + "/InvoiceDocumentSignature.cshtml");
+			string template = File.ReadAllText(_documentLocationRoot + "/InvoiceDocumentSignature.cshtml");
+			_logger.LogInformation("I think i got it");
+			string result = Engine.Razor.RunCompile(template, DateTime.Now.ToString(), null, model);
+			// instantiate the html to pdf converter
+			HtmlToPdf converter = new();
+			converter.Options.MarginBottom = 30;
+			converter.Options.MarginTop = 30;
+			// convert the url to pdf
+			PdfDocument doc = converter.ConvertHtmlString(result);
+
+			// save pdf document
+			var res = doc.Save();
+
+			// close pdf document
+			return res;
+		}
+
+		private string CalculateInvoicePeriod(InvoiceEmailModel model)
+		{
+			// Calculate the invoicing period
 			List<string> tempMonth = new();
 			List<string> tempYear = new();
-			foreach (var item in output.InvoiceSessions)
+			foreach (var item in model.InvoiceSessions)
 			{
 				tempMonth.Add(item.Date.ToString("MMMM"));
 				tempYear.Add(item.Date.Year.ToString());
@@ -87,10 +205,9 @@ namespace RevXApi.Library.DataAccess
 			var monthString = String.Join(" - ", tempMonth.Distinct());
 			var yearString = String.Join(" - ", tempYear.Distinct());
 
-			output.InvoicePeriod = $"{monthString} {yearString}";
+			string period = $"{monthString} {yearString}";
 
-			return output;
-
+			return period;
 		}
 
 		private string Convert24HourTo12Hour(string twentyfour)
@@ -99,21 +216,21 @@ namespace RevXApi.Library.DataAccess
 			var split = twentyfour.Split(":");
 			if (split?.Length == 3)
 			{
-				if (int.Parse(split[0]) == 12 )
+				if (int.Parse(split[ 0 ]) == 12)
 				{
-					output = $"12:{split[1]} PM";
+					output = $"12:{split[ 1 ]} PM";
 				}
-				else if (int.Parse(split[0]) == 0)
+				else if (int.Parse(split[ 0 ]) == 0)
 				{
-					output = $"12:{split[1]} AM";
+					output = $"12:{split[ 1 ]} AM";
 				}
-				else if (int.Parse(split[0]) < 12)
+				else if (int.Parse(split[ 0 ]) < 12)
 				{
-					output = $"{int.Parse(split[0])}:{split[1]} AM";
+					output = $"{int.Parse(split[ 0 ])}:{split[ 1 ]} AM";
 				}
-				else if (int.Parse(split[0]) > 12)
+				else if (int.Parse(split[ 0 ]) > 12)
 				{
-					output = $"{int.Parse(split[0]) - 12}:{split[1]} PM";
+					output = $"{int.Parse(split[ 0 ]) - 12}:{split[ 1 ]} PM";
 				}
 			}
 			return output;
