@@ -3,8 +3,6 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace RevXApi.Library.DataAccess
 {
@@ -14,35 +12,43 @@ namespace RevXApi.Library.DataAccess
 		private readonly IStudentData _studentData;
 		private readonly IProviderData _providerData;
 		private readonly IBillingStatusData _billingStatusData;
+		private readonly IHourlyRateData _rateData;
 
-		public SessionData(ISqlDataAccess sql, IStudentData studentData, IProviderData providerData, IBillingStatusData billingStatusData)
+		public SessionData(ISqlDataAccess sql, IStudentData studentData, IProviderData providerData, IBillingStatusData billingStatusData, IHourlyRateData rateData)
 		{
 			_sql = sql;
 			_studentData = studentData;
 			_providerData = providerData;
 			_billingStatusData = billingStatusData;
+			_rateData = rateData;
 		}
 
-		public List<SessionModel> GetAllSessions()
+		public List<SessionModel> GetAllSessions(string userId)
 		{
-			List<SessionDbModel> sessions = _sql.LoadData<SessionDbModel>("dbo.spSession_GetAll", "RevXData");
+			List<SessionDbModel> sessions = _sql.LoadData<SessionDbModel, dynamic>("dbo.spSession_GetAll", new { userId }, "RevXData");
 
 			List<SessionModel> output = new();
 
 			foreach (var session in sessions)
 			{
 				SessionModel model = new()
-				{ 
+				{
 					Id = session.Id,
+					UserId = session.UserId,
 					Date = session.Date,
 					StartTime = session.StartTime.ToString(),
 					EndTime = session.EndTime.ToString(),
 					Notes = session.Notes
 				};
 
-				model.Student = _studentData.GetById(session.StudentId);
-				model.Provider = _providerData.GetById(session.ProviderId);
+				model.Student = _studentData.GetById(session.StudentId, session.UserId);
+				model.Provider = _providerData.GetById(session.ProviderId, session.UserId);
 				model.BillingStatus = _billingStatusData.GetById(session.BillingStatusId);
+				model.Rate = _rateData.GetByDate(session.Date, session.UserId, session.ProviderId);
+				if (model.Rate == null)
+				{
+					model.Rate = new HourlyRate();
+				}
 
 
 				output.Add(model);
@@ -51,22 +57,28 @@ namespace RevXApi.Library.DataAccess
 			return output;
 		}
 
-		public SessionModel GetById(int id)
+		public SessionModel GetById(int id, string userId)
 		{
-			SessionDbModel session = _sql.LoadData<SessionDbModel, dynamic>("dbo.spSession_GetById", new { Id = id}, "RevXData").FirstOrDefault();
+			SessionDbModel session = _sql.LoadData<SessionDbModel, dynamic>("dbo.spSession_GetById", new { Id = id, userId }, "RevXData").FirstOrDefault();
 
 			SessionModel output = new()
 			{
 				Id = session.Id,
+				UserId = session.UserId,
 				Date = session.Date,
 				StartTime = session.StartTime.ToString(),
 				EndTime = session.EndTime.ToString(),
 				Notes = session.Notes
 			};
 
-			output.Student = _studentData.GetById(session.StudentId);
-			output.Provider = _providerData.GetById(session.ProviderId);
+			output.Student = _studentData.GetById(session.StudentId, session.UserId);
+			output.Provider = _providerData.GetById(session.ProviderId, session.UserId);
 			output.BillingStatus = _billingStatusData.GetById(session.BillingStatusId);
+			output.Rate = _rateData.GetByDate(session.Date, session.UserId, session.ProviderId);
+			if (output.Rate == null)
+{
+				output.Rate = new HourlyRate();
+			}
 
 			return output;
 		}
@@ -84,15 +96,21 @@ namespace RevXApi.Library.DataAccess
 				SessionModel model = new()
 				{
 					Id = session.Id,
+					UserId = session.UserId,
 					Date = session.Date,
 					StartTime = session.StartTime.ToString(),
 					EndTime = session.EndTime.ToString(),
 					Notes = session.Notes
 				};
 
-				model.Student = _studentData.GetById(session.StudentId);
-				model.Provider = _providerData.GetById(session.ProviderId);
+				model.Student = _studentData.GetById(session.StudentId, session.UserId);
+				model.Provider = _providerData.GetById(session.ProviderId, session.UserId);
 				model.BillingStatus = _billingStatusData.GetById(session.BillingStatusId);
+				model.Rate = _rateData.GetByDate(session.Date, session.UserId, session.ProviderId);
+				if (model.Rate == null)
+				{
+					model.Rate = new HourlyRate();
+				}
 
 				output.Add(model);
 			}
@@ -100,11 +118,12 @@ namespace RevXApi.Library.DataAccess
 			return output;
 		}
 
-		public void SaveSession(SessionModel model)
+		public int SaveSession(SessionModel model)
 		{
 			var dbModel = new SessionDbModel()
 			{
 				StudentId = model.Student.Id,
+				UserId = model.UserId,
 				Date = model.Date,
 				StartTime = ConvertToTimeSpan(model.StartTime),
 				EndTime = ConvertToTimeSpan(model.EndTime),
@@ -113,7 +132,7 @@ namespace RevXApi.Library.DataAccess
 				Notes = model.Notes
 			};
 
-			_sql.SaveData("dbo.spSession_Insert", dbModel, "RevXData");
+			return _sql.SaveData("dbo.spSession_Insert", dbModel, "RevXData");
 		}
 
 		public void EditSession(SessionModel model)
@@ -121,6 +140,7 @@ namespace RevXApi.Library.DataAccess
 			var dbModel = new SessionDbModel()
 			{
 				Id = model.Id,
+				UserId = model.UserId,
 				StudentId = model.Student.Id,
 				Date = model.Date,
 				StartTime = ConvertToTimeSpan(model.StartTime),
@@ -133,16 +153,16 @@ namespace RevXApi.Library.DataAccess
 			_sql.SaveData("dbo.spSession_Edit", dbModel, "RevXData");
 		}
 
-		public void DeleteSession(int id)
+		public void DeleteSession(int id, string userId)
 		{
-			_sql.SaveData("dbo.spSession_Delete", new { Id = id }, "RevXData");
+			_sql.SaveData("dbo.spSession_Delete", new { Id = id, userId }, "RevXData");
 		}
 
 		private TimeSpan ConvertToTimeSpan(string timeString)
 		{
 			var split = timeString.Split(':');
-			int.TryParse(split[0], out int hours);
-			int.TryParse(split[1], out int minutes);
+			int.TryParse(split[ 0 ], out int hours);
+			int.TryParse(split[ 1 ], out int minutes);
 			var output = new TimeSpan(hours, minutes, 00);
 			return output;
 		}
