@@ -1,9 +1,6 @@
 ﻿using RevXApi.Library.Models;
-using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace RevXApi.Library.DataAccess
 {
@@ -16,20 +13,30 @@ namespace RevXApi.Library.DataAccess
 			_sql = sql;
 		}
 
-		public List<StudentModel> GetAll()
+		public List<StudentModel> GetAll(string userId)
 		{
-			return _sql.LoadData<StudentModel>("dbo.spStudent_GetAll", "RevXData");
+			return _sql.LoadData<StudentModel, dynamic>("dbo.spStudent_GetAll", new { userId }, "RevXData");
 		}
 
-		public StudentModel GetById(int id)
+		public List<StudentModel> GetEnabled(string userId)
 		{
-			return _sql.LoadData<StudentModel, dynamic>("dbo.spStudent_GetById", new { Id = id }, "RevXData").FirstOrDefault();
-			
+			return _sql.LoadData<StudentModel, dynamic>("dbo.spStudent_GetEnabled", new { userId }, "RevXData");
+		}
+
+		public StudentModel GetById(int id, string userId)
+		{
+			return _sql.LoadData<StudentModel, dynamic>("dbo.spStudent_GetById", new { Id = id, userId }, "RevXData").FirstOrDefault();
+
 		}
 
 		public void AddStudent(StudentModel model)
 		{
 			_sql.SaveData("dbo.spStudent_Insert", model, "RevXData");
+		}
+
+		public void EditStudent(StudentModel model)
+		{
+			_sql.SaveData("dbo.spStudent_Update", model, "RevXData");
 		}
 	}
 }
