@@ -1,15 +1,9 @@
-﻿
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace RevXApi.Library.Models
+﻿namespace RevXApi.Library.Models
 {
 	public class BillingStatusModel
 	{
 		public int Id { get; set; }
 		public string BillingStatus { get; set; }
+		public bool Enabled { get; set; }
 	}
 }
