@@ -1,14 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace RevXApi.Library.Models
+﻿namespace RevXApi.Library.Models
 {
 	public class InvoiceDetailModel
 	{
 		public int InvoiceId { get; set; }
+		public string UserId { get; set; }
 		public int SessionId { get; set; }
 	}
 }
