@@ -1,14 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace RevXApi.Library.Models
+﻿namespace RevXApi.Library.Models
 {
 	public class ProviderModel
 	{
 		public int Id { get; set; }
+		public string UserId { get; set; }
 		public string Name { get; set; }
+		public bool Enabled { get; set; }
 	}
 }
