@@ -1,15 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace RevXApi.Library.Models
+﻿namespace RevXApi.Library.Models
 {
 	public class StudentModel
 	{
 		public int Id { get; set; }
+		public string UserId { get; set; }
 		public string FirstName { get; set; }
 		public string LastName { get; set; }
+		public bool Enabled { get; set; }
 	}
 }
