@@ -1,12 +1,9 @@
 ﻿using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using RevXApi.Library.DataAccess;
 using RevXApi.Library.Models;
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+using System.Security.Claims;
 
 namespace RevXApi.Controllers
 {
@@ -25,19 +22,33 @@ namespace RevXApi.Controllers
 		[HttpGet]
 		public List<StudentModel> GetAllStudents()
 		{
-			return _studentData.GetAll();
+			return _studentData.GetAll(User.FindFirstValue(ClaimTypes.NameIdentifier));
 		}
 
-		[HttpPost]
+		[HttpGet("enabled")]
+		public List<StudentModel> GetEnabled()
+		{
+			return _studentData.GetEnabled(User.FindFirstValue(ClaimTypes.NameIdentifier));
+		}
+
+		[HttpPost("add")]
 		public void AddStudent(StudentModel model)
 		{
+			model.UserId = User.FindFirstValue(ClaimTypes.NameIdentifier);
 			_studentData.AddStudent(model);
+		}
+
+		[HttpPost("edit")]
+		public void EditStudent(StudentModel model)
+		{
+			model.UserId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+			_studentData.EditStudent(model);
 		}
 
 		[HttpGet("{id}")]
 		public StudentModel GetStudentById(int id)
 		{
-			return _studentData.GetById(id);
+			return _studentData.GetById(id, User.FindFirstValue(ClaimTypes.NameIdentifier));
 		}
 	}
 }
