@@ -2,6 +2,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
+using NLog.Web;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -20,7 +21,13 @@ namespace RevXApi
 			Host.CreateDefaultBuilder(args)
 				.ConfigureWebHostDefaults(webBuilder =>
 				{
-					webBuilder.UseStartup<Startup>();
+					webBuilder.UseStartup<Startup>()
+					 .ConfigureLogging(logging =>
+					 {
+						 logging.ClearProviders();
+						 logging.SetMinimumLevel(LogLevel.Trace);
+					 })
+					.UseNLog();  // NLog: setup NLog for Dependency injection;
 				});
 	}
 }
