@@ -1,16 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace RevXPortal.Models
+﻿namespace RevXPortal.Models
 {
-	public class BillingStatusModel : ISelectionFriendly
+	public class BillingStatusModel : ISelectionFriendly, IEnable
 	{
 		public int Id { get; set; }
 		public string BillingStatus { get; set; }
+		public bool Enabled { get; set; }
 
 		public string DisplayName => BillingStatus;
+		public bool isEditMode { get; set; }
 	}
 }
