@@ -1,12 +1,6 @@
-﻿using RevXPortal.Converters;
-using RevXPortal.FormValidationAttributes;
+﻿using RevXPortal.FormValidationAttributes;
 using System;
-using System.Collections.Generic;
-using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
-using System.Diagnostics.CodeAnalysis;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace RevXPortal.Models
 {
@@ -15,6 +9,7 @@ namespace RevXPortal.Models
 	public class ManageSessionModel
 	{
 		public int Id { get; set; }
+		public string UserId { get; set; }
 		[Required]
 		public StudentModel Student { get; set; }
 		[Required]
@@ -28,7 +23,8 @@ namespace RevXPortal.Models
 		public ProviderModel Provider { get; set; }
 		[Required]
 		public BillingStatusModel BillingStatus { get; set; }
-		
+
 		public string Notes { get; set; }
+		public HourlyRate Rate { get; set; }
 	}
 }
