@@ -1,13 +1,11 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace RevXPortal.Models
 {
 	public class SessionApiModel
 	{
 		public int Id { get; set; }
+		public string UserId { get; set; }
 		public StudentModel Student { get; set; }
 		public DateTime Date { get; set; }
 		public string StartTime { get; set; }
@@ -15,5 +13,7 @@ namespace RevXPortal.Models
 		public ProviderModel Provider { get; set; }
 		public BillingStatusModel BillingStatus { get; set; }
 		public string Notes { get; set; }
+		public HourlyRate Rate { get; set; }
 	}
+
 }
