@@ -1,4 +1,4 @@
-﻿@page "/"
+﻿@page "/bla"
 
 <div style="text-align: center; margin-top: 6rem;">
     <h1>Welcome To RevX!!!</h1>
