@@ -1,59 +1,53 @@
-﻿using RevXPortal.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using System.Timers;
+﻿using System.Timers;
 
 namespace RevXPortal.Services
 {
+	public class ToastService : IDisposable, IToastService
+	{
+		public event Action<string, ToastLevel> OnShow;
+		public event Action OnHide;
+		private System.Timers.Timer Countdown;
 
-    public class ToastService : IDisposable
-    {
-        public event Action<string, ToastLevel> OnShow;
-        public event Action OnHide;
-        private Timer Countdown;
+		public void ShowToast(string message, ToastLevel level)
+		{
+			OnShow?.Invoke(message, level);
+			StartCountdown();
+		}
 
-        public void ShowToast(string message, ToastLevel level)
-        {
-            OnShow?.Invoke(message, level);
-            StartCountdown();
-        }
+		private void StartCountdown()
+		{
+			SetCountdown();
 
-        private void StartCountdown()
-        {
-            SetCountdown();
+			if (Countdown.Enabled)
+			{
+				Countdown.Stop();
+				Countdown.Start();
+			}
+			else
+			{
+				Countdown.Start();
+			}
+		}
 
-            if (Countdown.Enabled)
-            {
-                Countdown.Stop();
-                Countdown.Start();
-            }
-            else
-            {
-                Countdown.Start();
-            }
-        }
+		private void SetCountdown()
+		{
+			if (Countdown == null)
+			{
+				Countdown = new System.Timers.Timer(5000);
+				Countdown.Elapsed += HideToast;
+				Countdown.AutoReset = false;
+			}
+		}
 
-        private void SetCountdown()
-        {
-            if (Countdown == null)
-            {
-                Countdown = new Timer(5000);
-                Countdown.Elapsed += HideToast;
-                Countdown.AutoReset = false;
-            }
-        }
+		private void HideToast(object source, ElapsedEventArgs args)
+		{
+			OnHide?.Invoke();
+		}
 
-        private void HideToast(object source, ElapsedEventArgs args)
-        {
-            OnHide?.Invoke();
-        }
-
-        public void Dispose()
-        {
-            Countdown?.Dispose();
-        }
-    }
+		public void Dispose()
+		{
+			Countdown?.Dispose();
+		}
+	}
 
 }
