@@ -1,4 +1,4 @@
-﻿@inject RevXPortal.Services.ToastService toastService
+﻿@inject RevXPortal.Services.IToastService toastService
 <style>
     .toast {
         display: none;
