@@ -1,19 +1,31 @@
 ﻿@using System.Net.Http
-@using System.Linq 
+@using System.Linq
 @using System.Net.Http.Json
 @using System.Security.Claims
+@*@using Newtonsoft.Json*@
+@using System.Text.Json
 @using Microsoft.AspNetCore.Authorization
 @using Microsoft.AspNetCore.Components.Authorization
 @using Microsoft.AspNetCore.Components.Forms
 @using Microsoft.AspNetCore.Components.Routing
 @using Microsoft.AspNetCore.Components.Web
 @using Microsoft.AspNetCore.Components.Web.Virtualization
+@using System.Text.RegularExpressions;
 @using Microsoft.AspNetCore.Components.WebAssembly.Http
 @using Microsoft.JSInterop
 @using RevXPortal
 @using RevXPortal.Shared
+@using RevXPortal.Shared.InvoiceComponents
+@using RevXPortal.Shared.Tabs
+@using RevXPortal.Shared.Accordion
+@using RevXPortal.Shared.Calendar
+@using RevXPortal.Shared.Download
+@using RevXPortal.Converters
 @using RevXPortal.Authentication
 @using RevXPortal.Models
-@using RevXPortal.API 
-@using RevXPortal.Services 
-
+@using RevXPortal.API
+@using RevXPortal.Services
+@using System.IO
+@using OfficeOpenXml
+@using Microsoft.AspNetCore.WebUtilities
+@using Microsoft.Extensions.Primitives
